+++ R/plots.R
@@ -70,9 +70,8 @@
 #' class(popTimeData)
 #' popTimeData <- popTime(data = bmtcrr, time = "ftime", exposure = "D")
 #' attr(popTimeData, "exposure")
-#' @import data.table
 #' @export
-#' @importFrom data.table as.data.table rbindlist
+#' @importFrom data.table as.data.table rbindlist := setnames .N
 #' @importFrom stats quantile
 popTime <- function(data, time, event, censored.indicator,
                     exposure, percentile_number) {
@@ -282,11 +281,13 @@ popTime <- function(data, time, event, censored.indicator,
 
 # taken verbatim from cowplot::theme_cowplot()
 #' @importFrom stats quantile
+#' @importFrom grid unit
+#' @importFrom ggplot2 theme_grey theme element_line element_rect element_text margin element_blank rel %+replace%
 theme_cb <- function (font_size = 14, font_family = "", line_size = 0.5,
                       rel_small = 12/14, rel_tiny = 11/14, rel_large = 16/14) {
   half_line <- font_size/2
   small_size <- rel_small * font_size
-  theme_grey(base_size = font_size, base_family = font_family) %+replace%
+  ggplot2::theme_grey(base_size = font_size, base_family = font_family) %+replace%
     theme(line = element_line(color = "black", size = line_size,
                               linetype = 1, lineend = "butt"), rect = element_rect(fill = NA,
                                                                                    color = NA, size = line_size, linetype = 1), text = element_text(family = font_family,
@@ -344,6 +345,7 @@ theme_cb <- function (font_size = 14, font_family = "", line_size = 0.5,
 
 
 # taken verbatim from cowplot::panel_border
+#' @importFrom ggplot2 theme element_blank element_rect
 panelBorder <- function(color = "grey85", size = 1, linetype = 1, remove = FALSE,
                         colour) {
   if (!missing(colour)) {
@@ -363,7 +365,7 @@ panelBorder <- function(color = "grey85", size = 1, linetype = 1, remove = FALSE
 #' @rdname plot.singleEventCB
 #' @param newdata2 `data.frame` for exposed group. calculated and passed
 #'   internally to `plotHazardRatio` function
-#' @importFrom graphics polygon
+#' @importFrom graphics polygon lines arrows points
 #' @importFrom stats terms delete.response vcov qnorm
 #' @importFrom utils modifyList
 plotHazardRatio <- function(x, newdata, newdata2, ci, ci.lvl, ci.col,
@@ -397,6 +399,15 @@ plotHazardRatio <- function(x, newdata, newdata2, ci, ci.lvl, ci.col,
   i.backw <- order(xvar_values, decreasing = TRUE)
   i.forw <- order(xvar_values)
 
+  other_plot_args <- list(...)
+  line_args <- grep("^lwd$|^lty$|^col$|^pch$|^cex$", names(other_plot_args))
+
+  if (length(line_args) == 0) {
+    line_args <- list(NULL)
+  } else {
+    line_args <- other_plot_args[line_args]
+  }
+
   # plot CI as polygon shade - if 'se = TRUE' (default)
   if (ci) {
     v2 <- stats::vcov(x)
@@ -407,20 +418,10 @@ plotHazardRatio <- function(x, newdata, newdata2, ci, ci.lvl, ci.col,
     x.poly <- c(xvar_values[i.forw], xvar_values[i.backw])
     y.poly <- c(hazard_ratio_lower[i.forw], hazard_ratio_upper[i.backw])
 
-    other_plot_args <- list(...)
-
-    line_args <- grep("^lwd$|^lty$|^col$", names(other_plot_args))
-
-    if (length(line_args) == 0) {
-      line_args <- list(NULL)
-    } else {
-      line_args <- other_plot_args[line_args]
-    }
-
     do.call("plot", utils::modifyList(
       list(
         x = range(x.poly),
-        y = range(y.poly),
+        y = range(y.poly)*c(0.99, 1.01),
         type = "n",
         ylab = "hazard ratio",
         xlab = xvar
@@ -428,20 +429,31 @@ plotHazardRatio <- function(x, newdata, newdata2, ci, ci.lvl, ci.col,
       other_plot_args
     ))
 
-    graphics::polygon(x.poly, y.poly, col = ci.col, border = NA)
-
-    do.call("lines", utils::modifyList(
-      list(
-        x = xvar_values[i.forw],
-        y = exp(log_hazard_ratio[i.forw]),
-        lwd = 2,
-        lty = 1,
-        col = "black"
-      ),
-      line_args
-    ))
-
-    # browser()
+    if (length(unique(x.poly)) == 1) {
+      graphics::arrows(x0 = x.poly[1], y0 = y.poly[1], y1 = y.poly[2], angle = 90, length = 0.5, code = 3)
+      do.call("points", utils::modifyList(
+        list(
+          x = xvar_values[i.forw],
+          y = exp(log_hazard_ratio[i.forw]),
+          pch = 19,
+          col = "black",
+          cex = 2
+        ),
+        line_args
+      ))
+    } else {
+      graphics::polygon(x.poly, y.poly, col = ci.col, border = NA)
+      do.call("lines", utils::modifyList(
+        list(
+          x = xvar_values[i.forw],
+          y = exp(log_hazard_ratio[i.forw]),
+          lwd = 2,
+          lty = 1,
+          col = "black"
+        ),
+        line_args
+      ))
+    }
 
     results <- transform(newdata,
       log_hazard_ratio = log_hazard_ratio,
@@ -450,14 +462,30 @@ plotHazardRatio <- function(x, newdata, newdata2, ci, ci.lvl, ci.col,
       lowerbound = hazard_ratio_lower,
       upperbound = hazard_ratio_upper
     )
+
   } else {
-    do.call("plot", utils::modifyList(
-      list(
-        x = xvar_values, y = exp(log_hazard_ratio), lwd = 2, lty = 1, type = "l",
-        ylab = "hazard ratio", xlab = xvar
-      ),
-      list(...)
-    ))
+
+    # browser()
+
+    if (length(xvar_values) == 1) {
+      do.call("plot", utils::modifyList(
+        list(
+          x = xvar_values, y = exp(log_hazard_ratio), lwd = 2, lty = 1,
+          pch = 19, cex = 2, ylab = "hazard ratio", xlab = xvar
+        ),
+        other_plot_args
+      ))
+
+    } else {
+
+      do.call("plot", utils::modifyList(
+        list(
+          x = xvar_values, y = exp(log_hazard_ratio), lwd = 2, lty = 1, type = "l",
+          ylab = "hazard ratio", xlab = xvar
+        ),
+        other_plot_args
+      ))
+    }
 
     results <- transform(newdata,
       log_hazard_ratio = log_hazard_ratio,