+++ NEWS.md
@@ -1,4 +1,4 @@
-# casebase 1.0.0
+# casebase 0.9.0
 
 * *Major new release*
 * Some minor bug fixes
@@ -7,7 +7,10 @@
 * Users now have further control on the output of `absoluteRisk` using the arguments `type` and `addZero`.
 * The output of `absoluteRisk` now always contains the time variable in the first column, regardless of the length of `time`. This will break earlier code that depended on the previous behaviour.
 * Population time plots now use `geom_ribbon` instead of `geom_segment`
-* Population time functions now allowed for more flexible plots with user defined arguments
+* Population time functions now allowed for more flexible plots with user defined arguments including sequentially adding base, case, and competing event series
+* New plotting method for time-dependent hazard functions and hazard ratios. These include confidence intervals. See `plot.singleEventCB`. The hazard function plot requires `visreg` package to be installed. 
+* New plotting method for cumulative incidence and survival curves. See `plot.absRiskCB`.
+* No longer importing the entire namespace of `data.table` and `ggplot2`. 
 
 # casebase 0.1.0